requirements.txt

openai>=1.0.0
requests
selenium
google-api-python-client
google-auth
python-dotenv
